+++ src/main/java/se/lexicon/jpa_workshop/repository/DetailsRepository.java
@@ -12,7 +12,7 @@ import java.util.List;
 // and we can easily create custom version of those or use @Query to make more complex methods directly in the interface.
 // We don't need to implement these methods in a class, Spring data JPA will do it all for us.
 // We just need to use this interface as an object.
-public interface DetailsRepository extends CrudRepository<Details, Integer> {
+public interface DetailsRepository extends CrudRepository<Details, Long> {
     // By following Spring data JPA naming convention we can create custom version of CrudRepository's already existing methods.
     List<Details> findByEmail(String email);
     List<Details> findByEmailIgnoreCase(String email);
